# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Chronicle API
  version: 1.0.0
  description: REST API for Chronicle, a self-hosted TTRPG worldbuilding platform. Provides programmatic access to campaign data for external tool integration (Foundry VTT, custom scripts).
servers:
  - url: "{baseUrl}/api/v1"
    variables:
      baseUrl:
        default: http://localhost:8080
        description: Chronicle server base URL
security:
  - BearerApiKey: []
...
paths:
...
  /campaigns/{campaignId}/maps:
    get:
      tags: [Maps]
      summary: List maps
      operationId: listMaps
      description: Requires maps addon to be enabled.
      parameters:
        - $ref: "#/components/parameters/campaignId"
      responses:
        "200":
          description: List of maps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Map"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /campaigns/{campaignId}/maps/{mapId}:
    get:
      tags: [Maps]
      summary: Get map details
      operationId: getMap
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
      responses:
        "200":
          description: Map with markers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Map"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/maps/{mapId}/drawings:
    get:
      tags: [Maps]
      summary: List drawings
      operationId: listDrawings
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
      responses:
        "200":
          description: List of drawings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Drawing"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Maps]
      summary: Create drawing
      operationId: createDrawing
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDrawingRequest"
      responses:
        "201":
          description: Created drawing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drawing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/maps/{mapId}/drawings/{drawingId}:
    put:
      tags: [Maps]
      summary: Update drawing
      operationId: updateDrawing
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
        - $ref: "#/components/parameters/drawingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDrawingRequest"
      responses:
        "204":
          description: Drawing updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      tags: [Maps]
      summary: Delete drawing
      operationId: deleteDrawing
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
        - $ref: "#/components/parameters/drawingId"
      responses:
        "204":
          description: Drawing deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/maps/{mapId}/tokens:
    get:
      tags: [Maps]
      summary: List tokens
      operationId: listTokens
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
      responses:
        "200":
          description: List of tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Maps]
      summary: Create token
      operationId: createToken
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "201":
          description: Created token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/maps/{mapId}/tokens/{tokenId}:
    put:
      tags: [Maps]
      summary: Update token
      operationId: updateToken
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
        - $ref: "#/components/parameters/tokenId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTokenRequest"
      responses:
        "204":
          description: Token updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      tags: [Maps]
      summary: Delete token
      operationId: deleteToken
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
        - $ref: "#/components/parameters/tokenId"
      responses:
        "204":
          description: Token deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/maps/{mapId}/tokens/{tokenId}/position:
    patch:
      tags: [Maps]
      summary: Update token position
      operationId: updateTokenPosition
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
        - $ref: "#/components/parameters/tokenId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [x, y]
              properties:
                x:
                  type: number
                  format: double
                y:
                  type: number
                  format: double
      responses:
        "204":
          description: Position updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/maps/{mapId}/layers:
    get:
      tags: [Maps]
      summary: List layers
      operationId: listLayers
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
      responses:
        "200":
          description: List of layers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Layer"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Maps]
      summary: Create layer
      operationId: createLayer
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLayerRequest"
      responses:
        "201":
          description: Created layer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Layer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/maps/{mapId}/layers/{layerId}:
    put:
      tags: [Maps]
      summary: Update layer
      operationId: updateLayer
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
        - $ref: "#/components/parameters/layerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLayerRequest"
      responses:
        "204":
          description: Layer updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      tags: [Maps]
      summary: Delete layer
      operationId: deleteLayer
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
        - $ref: "#/components/parameters/layerId"
      responses:
        "204":
          description: Layer deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/maps/{mapId}/fog:
    get:
      tags: [Maps]
      summary: List fog regions
      operationId: listFog
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
      responses:
        "200":
          description: List of fog regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FogRegion"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Maps]
      summary: Create fog region
      operationId: createFog
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFogRequest"
      responses:
        "201":
          description: Created fog region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FogRegion"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      tags: [Maps]
      summary: Reset all fog
      operationId: resetFog
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
      responses:
        "204":
          description: All fog cleared
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/maps/{mapId}/fog/{fogId}:
    delete:
      tags: [Maps]
      summary: Delete fog region
      operationId: deleteFog
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/mapId"
        - $ref: "#/components/parameters/fogId"
      responses:
        "204":
          description: Fog region deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

  # --- Sync ---
  /campaigns/{campaignId}/sync:
    post:
      tags: [Sync]
      summary: Batch sync (push/pull changes)
      operationId: sync
      description: >
        Bidirectional sync endpoint. Provide "since" to pull entities modified
        after that timestamp. Provide "changes" array to push create/update/delete
        operations. Max 2000 changes per request.
      parameters:
        - $ref: "#/components/parameters/campaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
      responses:
        "200":
          description: Sync results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: >
        API key passed as Bearer token. Keys are scoped to a single campaign
        with permission levels: read, write, sync. Rate limited per key.
  parameters:
    campaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
...
    mapId:
      name: mapId
      in: path
      required: true
      schema:
        type: string
    drawingId:
      name: drawingId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Map:
      type: object
      properties:
        id:
          type: string
        campaign_id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        image_id:
          type: string
          nullable: true
        image_width:
          type: integer
        image_height:
          type: integer
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        markers:
          type: array
          items:
            $ref: "#/components/schemas/Marker"
...